add: 日本語と英語で共通しているものをコンポーネント化して、共通化(でも冗長かも)
